refactor: 로딩 페이지 로직 분리 및 컴포넌트에 memo 사용
